Data Layer Cleanup
A lot of repackaging, renaming, and altering models for better clarity.
