+++ android/app/src/main/java/com/emergetools/hackernews/features/bookmarks/BookmarksRouting.kt
@@ -1,17 +1,15 @@
 package com.emergetools.hackernews.features.bookmarks
 
 import android.net.Uri
-import androidx.browser.customtabs.CustomTabsIntent
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import com.emergetools.hackernews.LocalCustomTabsIntent
 import com.emergetools.hackernews.bookmarkDao
-import com.emergetools.hackernews.data.LocalCustomTabsIntent
 import com.emergetools.hackernews.features.bookmarks.BookmarksDestinations.Bookmarks
 import kotlinx.serialization.Serializable
