+++ android/app/src/main/java/com/emergetools/hackernews/features/settings/SettingsDomain.kt
@@ -3,14 +3,13 @@ package com.emergetools.hackernews.features.settings
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import com.emergetools.hackernews.data.UserStorage
+import com.emergetools.hackernews.data.local.UserStorage
 import com.emergetools.hackernews.features.login.LoginDestinations
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
 data class SettingsState(
@@ -19,13 +18,14 @@ data class SettingsState(
 
 sealed interface SettingsAction {
   data object LoginPressed : SettingsAction
-  data object LogoutPressed: SettingsAction
+  data object LogoutPressed : SettingsAction
 }
 
 sealed interface SettingsNavigation {
   data object GoToLogin : SettingsNavigation {
     val login = LoginDestinations.Login
   }
+
   data class GoToSettingsLink(val url: String) : SettingsNavigation
 }
 
@@ -57,7 +57,7 @@ class SettingsViewModel(private val userStorage: UserStorage) : ViewModel() {
   }
 
   @Suppress("UNCHECKED_CAST")
-  class Factory(private val userStorage: UserStorage): ViewModelProvider.Factory {
+  class Factory(private val userStorage: UserStorage) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
       return SettingsViewModel(userStorage) as T
     }
