+++ android/app/src/main/java/com/emergetools/hackernews/features/stories/StoriesRouting.kt
@@ -1,11 +1,8 @@
 package com.emergetools.hackernews.features.stories
 
 import android.net.Uri
-import androidx.browser.customtabs.CustomTabsIntent
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
@@ -13,11 +10,11 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
 import androidx.navigation.toRoute
+import com.emergetools.hackernews.LocalCustomTabsIntent
+import com.emergetools.hackernews.baseClient
 import com.emergetools.hackernews.bookmarkDao
-import com.emergetools.hackernews.data.LocalCustomTabsIntent
 import com.emergetools.hackernews.features.stories.StoriesDestinations.Closeup
 import com.emergetools.hackernews.features.stories.StoriesDestinations.Feed
-import com.emergetools.hackernews.itemRepository
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -39,7 +36,7 @@ fun NavGraphBuilder.storiesGraph(navController: NavController) {
 
       val model = viewModel<StoriesViewModel>(
         factory = StoriesViewModel.Factory(
-          itemRepository = context.itemRepository(),
+          baseClient = context.baseClient(),
           bookmarkDao = context.bookmarkDao()
         )
       )
